Give pli-extras higher priority than other layers
Adding a recipe to pli-extras with an existing name in another layer
will be ignored, as pli-extras has the lowest priority,
even when the version in pli-extras is higher.
Give pli-extras higher priority to fix that.

This triggers a rebuild of all packages, sorry.
Recommended to delete the entire 'build' directory.